services/stt-service/requirements.txt
# FastAPI + HTTP server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# gRPC
grpcio==1.68.1
grpcio-tools==1.68.1
grpcio-health-checking==1.68.1
protobuf==5.29.2

# Speech-to-Text
faster-whisper>=1.0.0
ctranslate2>=4.0.0

# Voice Activity Detection
torch>=2.1.0
torchaudio>=2.1.0

# Language Detection
fasttext-wheel>=0.9.2

# Audio processing
numpy>=1.26.0
soundfile==0.12.1

# Configuration
pydantic==2.10.3
pydantic-settings==2.7.0

# Observability
structlog==24.4.0
prometheus-client==0.21.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp==1.29.0

# Redis (for session events)
redis[hiredis]==5.2.1
